test: split image and reasonig out from base
trying to create a good base line of features that all providers would
support
