+++ test/fixtures/vcr_cassettes/integration/live/stream_image_test_rb/test_live_image_streaming_anthropic_apikey_messages_claude-sonnet-4-20250514.yml
