+++ test/fixtures/vcr_cassettes/integration/live/stream_image_test_rb/test_live_image_streaming_openai_apikey_completions_gpt-5_1.yml
