+++ test/fixtures/vcr_cassettes/integration/live/stream_image_test_rb/test_live_image_streaming_openai_oauth_codex_gpt-5_4.yml
