+++ test/fixtures/vcr_cassettes/integration/live/stream_reasoning_test_rb/test_live_basic_thinking_anthropic_oauth_messages_claude-sonnet-4-20250514.yml
