+++ test/fixtures/vcr_cassettes/integration/live/stream_reasoning_test_rb/test_live_basic_thinking_openai_oauth_codex_gpt-5_4.yml
